ID-821 move fragments in another file
